+++ src/tls/mod.rs
@@ -38,7 +38,6 @@ mod session;
 mod types;
 
 pub use factory::{TlsFactory, TlsReassembler};
-pub use parser::Error;
 pub use session::{TlsMessage, TlsParser};
 pub use types::*;